Change time durations to int_64
For seconds best that can be done is unsigned long (4 bits) which is
about 136 years.  Other timescales may allow longer time periods.
Javascript variables are converted to strings then are converted to
floats and then converted to int64_t.  Limitations of snprintf in Pebble
prevent display of int64_t.  Best we can do is unsigned long.